fix: resolve circular imports and update OpenAI configuration
- Fix circular imports between cache.py and utils.py
- Update OpenAI API endpoint and model configuration
- Add openai package to requirements.txt
- Remove pre-compiled regex patterns to fix pattern compilation error
